wedding_bot/requirements.txt
python-telegram-bot[all]==21.0
aiosqlite==0.19.0
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.23
APScheduler==3.10.4
aiohttp==3.9.1
